requirements.txt
Flask==3.0.3
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.1
bcrypt==4.1.2
pg8000==1.31.2
